test(replace-remix-imports): fix `shell.grep` call
previously, `shell.grep` was being called with the `-r` option, which it does not recognize,
resulting in an exit ode of 1. we incorrectly interpreted that as `grep` finishing the search
without finding matches. instead, the `shell.grep` call was crashing. now we correctly pass a list
of files to `shell.grep` without the `-r` option.